+++ basis-universal/src/lib.rs
@@ -41,11 +41,11 @@ pub const ETC1S_QUALITY_MIN: u32 = sys::basisu_BASISU_QUALITY_MIN;
 pub const ETC1S_QUALITY_MAX: u32 = sys::basisu_BASISU_QUALITY_MAX;
 
 /// The default quality level used if [CompressorParams::set_uastc_quality_level] is not called
-pub const UASTC_QUALITY_DEFAULT: u32 = sys::UastcPackFlags_PackUASTCLevelDefault as u32;
+pub const UASTC_QUALITY_DEFAULT: u32 = sys::UastcPackFlags_PackUASTCLevelDefault /*as u32*/;
 /// The minimum quality level that can be provided to [CompressorParams::set_uastc_quality_level]
-pub const UASTC_QUALITY_MIN: u32 = sys::UastcPackFlags_PackUASTCLevelFastest as u32;
+pub const UASTC_QUALITY_MIN: u32 = sys::UastcPackFlags_PackUASTCLevelFastest /*as u32*/;
 /// The maximum quality level that can be provided to [CompressorParams::set_uastc_quality_level]
-pub const UASTC_QUALITY_MAX: u32 = sys::UastcPackFlags_PackUASTCLevelVerySlow as u32;
+pub const UASTC_QUALITY_MAX: u32 = sys::UastcPackFlags_PackUASTCLevelVerySlow /*as u32*/;
 
 /// Maximum supported texture dimension
 pub const TEXTURE_DIMENSION_MAX: u32 = sys::basisu_BASISU_MAX_SUPPORTED_TEXTURE_DIMENSION;
